[geometry] Move odometry into odometry namespace
