+++ ecl_geometry/include/ecl/CMakeLists.txt
@@ -3,12 +3,13 @@
 ###############################################################################
 
 add_subdirectory(geometry)
+add_subdirectory(odometry)
 
 ###############################################################################
 # FILES
 ###############################################################################
 
 file(GLOB HEADERS RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} *.hpp)
-  
+
 install(FILES ${HEADERS} DESTINATION ${CATKIN_GLOBAL_INCLUDE_DESTINATION}/ecl)
