+++ ecl_geometry/include/ecl/geometry.hpp
@@ -18,6 +18,7 @@
 *****************************************************************************/
 
 #include <ecl/linear_algebra.hpp>
+#include <ecl/odometry/odometry_typedefs.hpp>
 
 /*****************************************************************************
 ** Includes
