+++ ecl_geometry/include/ecl/odometry/CMakeLists.txt
@@ -1,0 +1,8 @@
+###############################################################################
+# FILES
+###############################################################################
+
+file(GLOB HEADERS RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} *.hpp)
+
+install(FILES ${HEADERS} DESTINATION ${CATKIN_GLOBAL_INCLUDE_DESTINATION}/ecl/odometry)
+
