+++ ecl_geometry/include/ecl/odometry/odometry.hpp
@@ -1,0 +1,7 @@
+#ifndef ECL_ODOMETRY_ODOMETRY_HPP_
+#define ECL_ODOMETRY_ODOMETRY_HPP_
+
+#include "odometry_typedefs.hpp"
+#include "odometry_helper.hpp"
+
+#endif /*ECL_ODOMETRY_ODOMETRY_HPP_*/
